requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.12
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
limits==4.2
mangum==0.17.0
openai==1.79.0
packaging==24.2
pyasn1==0.4.8
pydantic==2.11.4
pydantic-core==2.33.2
python-dotenv==1.1.0
python-jose==3.4.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
wrapt==1.17.2
